gitlab/idfa/requirements.txt
rq
redis
psycopg2-binary
click
gevent
rq-scheduler
